+++ lib/Pages/login.dart
@@ -1,20 +1,65 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_application_1/utils/routes.dart';
 
-class Login extends StatelessWidget {
+class login extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Material(
-      child: Center(
-        child: Text(
-          "Login Page",
-          style: TextStyle(
-            fontSize: 26,
-            color: Colors.blue,
-            fontWeight: FontWeight.bold,
+        child: SingleChildScrollView(
+      child: Column(
+        children: [
+          Image.asset(
+            "assets/images/undraw_Engineer.png",
+            fit: BoxFit.contain,
+            height: 500,
+            width: 1000,
           ),
-          textScaleFactor: 3.0,
-        ),
+          SizedBox(
+            height: 30.0,
+          ),
+          Text(
+            "Welcome",
+            style: TextStyle(
+              fontSize: 22,
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+          SizedBox(
+            height: 20.0,
+          ),
+          Padding(
+            padding:
+                const EdgeInsets.symmetric(vertical: 2.0, horizontal: 280.0),
+            child: Column(
+              children: [
+                TextFormField(
+                  decoration: InputDecoration(
+                    hintText: "Enter UserName",
+                    labelText: "User Name",
+                  ),
+                ),
+                TextFormField(
+                  obscureText: true,
+                  decoration: InputDecoration(
+                    hintText: "Enter Password",
+                    labelText: "Password",
+                  ),
+                ),
+                SizedBox(
+                  height: 30,
+                ),
+                ElevatedButton(
+                  onPressed: () {
+                    Navigator.pushNamed(context, MyRoutes.homeRoute);
+                  },
+                  child: Text("Login"),
+                  style: TextButton.styleFrom(),
+                )
+              ],
+            ),
+          ),
+        ],
       ),
-    );
+    ));
   }
 }
