+++ lib/utils/routes.dart
@@ -1,0 +1,4 @@
+class MyRoutes {
+  static String loginRoutes = "/login";
+  static String homeRoute = "/home";
+}
